move function from utils to pipeline to avoid circular import
